requirements.txt
###### Requirements without Version Specifiers ######
pcaspy
logging
re
pyepics
